+++ test/join_right.sql
@@ -1,4 +1,4 @@
-create table planets (name text, moons int4, primary key (name));
+create table planets (name text, moons int8, primary key (name));
 insert into planets (name, moons) values ('Earth', 1), ('Mars', 2), ('Venus', 0);
 
 create table moons (name text, planet text, primary key(name));
