+++ test/not_null_applied_for_primary_keys.sql
@@ -1,4 +1,4 @@
-create table planets(system text, name text, moons int4, primary key(system, name));
+create table planets(system text, name text, moons int8, primary key(system, name));
 
 describe table planets;
