+++ test/transaction_single_stmt.sql
@@ -1,4 +1,4 @@
-create table planets (id int4, name text, primary key (id));
+create table planets (id int8, name text, primary key (id));
 insert into planets (id, name) values (1, 'Earth'), (2, 'Mars');
 insert into planets (id, name) values (3, 'Jupiter'), (1, 'Saturn');
 insert into planets (id, name) values (1, 'Jupiter'), (4, 'Saturn');
